Improved CPU usage % (higher is better).

+++ ProbQA/SRPlatform/Interface/SRBasicTypes.h
@@ -8,7 +8,7 @@ namespace SRPlat {
 
 typedef double SRAmount;
 typedef uint32_t SRThreadCount;
-typedef int32_t SRSubtaskCount;
+typedef size_t SRSubtaskCount;
 typedef uint8_t SRVectCompCount;
 
 } // namespace SRPlat
